requirements.txt
Flask
mysql-connector-python
pandas
numpy
tensorflow
scikit-learn
joblib